access_controller_agent/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-adk
google-genai
requests
